Rename 'yaml.py' to 'loader.py'
