+++ xawscf/pattern/apigateway.py
@@ -1,5 +1,5 @@
 import yaml
-from ..utils.yaml import Custom
+from ..utils.loader import Custom
 from ..utils.helper import get_full_name, make_output
 from .base_resource import BaseResource
 from .role import Role
