+++ xawscf/pattern/function.py
@@ -1,5 +1,5 @@
 from ..utils import helper
-from ..utils.yaml import Custom
+from ..utils.loader import Custom
 from ..utils.const import SOURCES_BUCKET
 from ..utils.helper import get_full_name, try_set_field, make_output
 from .base_resource import BaseResource
